+++ cognite/experimental/_api/transformations.py
@@ -1,17 +1,17 @@
-import json as _json
-from typing import List, Optional, Union
+from typing import Awaitable, List, Optional, Union
 
 from cognite.client import utils
 from cognite.client._api_client import APIClient
-from requests import Response
 
+from cognite.experimental._api.transformation_jobs import TransformationJobsAPI
 from cognite.experimental._api.transformation_notifications import TransformationNotificationsAPI
 from cognite.experimental._api.transformation_schedules import TransformationSchedulesAPI
 from cognite.experimental._api.transformation_schema import TransformationSchemaAPI
-from cognite.experimental._constants import HANDLER_FILE_NAME, LIST_LIMIT_CEILING, LIST_LIMIT_DEFAULT, MAX_RETRIES
+from cognite.experimental._constants import LIST_LIMIT_CEILING, LIST_LIMIT_DEFAULT
 from cognite.experimental.data_classes import (
     Transformation,
     TransformationFilter,
+    TransformationJob,
     TransformationList,
     TransformationUpdate,
 )
@@ -23,7 +23,7 @@ class TransformationsAPI(APIClient):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        # self.jobs = TransformationJobsAPI(*args, **kwargs)
+        self.jobs = TransformationJobsAPI(*args, **kwargs)
         self.schedules = TransformationSchedulesAPI(*args, **kwargs)
         self.schema = TransformationSchemaAPI(*args, **kwargs)
         self.notifications = TransformationNotificationsAPI(*args, **kwargs)
@@ -48,7 +48,7 @@ class TransformationsAPI(APIClient):
                 >>> c = CogniteClient()
                 >>> transformations = [
                 >>>     Transformation(
-                >>>         name="transformation1", 
+                >>>         name="transformation1",
                 >>>         destination=TransformationDestination.assets()
                 >>>     ),
                 >>>     Transformation(
@@ -134,6 +134,26 @@ class TransformationsAPI(APIClient):
         utils._auxiliary.assert_exactly_one_of_id_or_external_id(id, external_id)
         return self._retrieve_multiple(ids=id, external_ids=external_id, wrap_ids=True)
 
+    def retrieve_multiple(self, ids: List[int] = None, external_ids: List[str] = None) -> TransformationList:
+        """`Retrieve multiple transformations. <https://docs.cognite.com/api/playground/#operation/getTransformation>`_
+
+        Args:
+            ids (List[int]): List of ids to retrieve
+            external_ids (List[str]): List of external ids to retrieve
+
+        Returns:
+            TransformationList: Requested transformation or None if it does not exist.
+
+        Examples:
+
+            Get multiple transformations:
+
+                >>> from cognite.experimental import CogniteClient
+                >>> c = CogniteClient()
+                >>> res = c.transformations.retrieve_multiple(ids=[1,2,3], external_ids=['transform-1','transform-2'])
+        """
+        return self._retrieve_multiple(ids=ids, external_ids=external_ids, wrap_ids=True)
+
     def update(
         self, item: Union[Transformation, TransformationUpdate, List[Union[Transformation, TransformationUpdate]]]
     ) -> Union[Transformation, TransformationList]:
@@ -164,3 +184,83 @@ class TransformationsAPI(APIClient):
                 >>> res = c.transformations.update(my_update)
         """
         return self._update_multiple(items=item)
+
+    def run(
+        self, transformation_id: int = None, transformation_external_id: str = None, wait: bool = True
+    ) -> TransformationJob:
+        """`Run a transformation. <https://docs.cognite.com/api/playground/#operation/runTransformation>`_
+
+        Args:
+            transformation_id (int): internal Transformation id
+            transformation_external_id (str): external Transformation id
+            wait (bool): Wait until the transformation run is finished. Defaults to True.
+
+        Returns:
+            Created transformation job
+
+        Examples:
+
+            Run transformation to completion by id:
+
+                >>> from cognite.experimental import CogniteClient
+                >>> c = CogniteClient()
+                >>>
+                >>> res = c.transformations.run(id = 1)
+
+            Start running transformation by id:
+
+                >>> from cognite.experimental import CogniteClient
+                >>> c = CogniteClient()
+                >>>
+                >>> res = c.transformations.run(id = 1, wait = False)
+        """
+        utils._auxiliary.assert_exactly_one_of_id_or_external_id(transformation_id, transformation_external_id)
+
+        if transformation_external_id:
+            transformation_id = self.retrieve(external_id=transformation_external_id).id
+
+        response = self._post(
+            url_path=utils._auxiliary.interpolate_and_url_encode(
+                self._RESOURCE_PATH + "/{}/run", str(transformation_id)
+            )
+        )
+        job = TransformationJob._load(response.json(), cognite_client=self._cognite_client)
+
+        if wait:
+            return job.wait()
+
+        return job
+
+    def run_async(
+        self, transformation_id: int = None, transformation_external_id: str = None
+    ) -> Awaitable[TransformationJob]:
+        """`Run a transformation to completion asynchronously. <https://docs.cognite.com/api/playground/#operation/runTransformation>`_
+
+        Args:
+            transformation_id (int): internal Transformation id
+            transformation_external_id (str): external Transformation id
+
+        Returns:
+            Completed transformation job
+
+        Examples:
+
+            Run transformation asyncronously by id:
+
+                >>> import asyncio
+                >>> from cognite.experimental import CogniteClient
+                >>>
+                >>> c = CogniteClient()
+                >>>
+                >>> async def run_transformation():
+                >>>     res = await c.transformations.run_async(id = 1)
+                >>>
+                >>> loop = asyncio.get_event_loop()
+                >>> loop.run_until_complete(run_transformation())
+                >>> loop.close()
+        """
+
+        job = self.run(
+            transformation_id=transformation_id, transformation_external_id=transformation_external_id, wait=False
+        )
+        return job.wait_async()