+++ tests/tests_integration/test_api/test_transformations.py
@@ -1,3 +1,6 @@
+import random
+import string
+
 import pytest
 
 from cognite.experimental import CogniteClient
@@ -8,13 +11,16 @@ COGNITE_CLIENT = CogniteClient()
 
 @pytest.fixture
 def new_transformation():
-    transform = Transformation(name="any", destination=TransformationDestination.raw())
+    transform = Transformation(name="any", destination=TransformationDestination.assets(),)
     ts = COGNITE_CLIENT.transformations.create(transform)
 
     yield ts
 
     COGNITE_CLIENT.transformations.delete(id=ts.id)
     assert COGNITE_CLIENT.transformations.retrieve(ts.id) is None
+
+
+other_transformation = new_transformation
 
 
 class TestTransformationsAPI:
@@ -36,8 +42,7 @@ class TestTransformationsAPI:
     def test_create(self, new_transformation):
         assert (
             new_transformation.name == "any"
-            and new_transformation.destination.type == "raw_table"
-            and new_transformation.destination.rawType == "plain_raw"
+            and new_transformation.destination.type == "assets"
             and new_transformation.id is not None
         )
 
@@ -48,6 +53,15 @@ class TestTransformationsAPI:
             and new_transformation.destination.type == retrieved_transformation.destination.type
             and new_transformation.id == retrieved_transformation.id
         )
+
+    def test_retrieve_multiple(self, new_transformation, other_transformation):
+        retrieved_transformations = COGNITE_CLIENT.transformations.retrieve_multiple(
+            ids=[new_transformation.id, other_transformation.id]
+        )
+        assert len(retrieved_transformations) == 2
+        assert new_transformation.id in [
+            transformation.id for transformation in retrieved_transformations
+        ] and other_transformation.id in [transformation.id for transformation in retrieved_transformations]
 
     def test_update_full(self, new_transformation):
         new_transformation.name = "new name"